backend/requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.4
sodapy==2.1.0
shapely==2.0.2
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
